Remove processError from Domain class
Wrap any fallible libvirt function in a call to ErrorHandler.processError(..).

Also update erroneous javadoc comments stating that methods would return a
value in case an error occurs.  In case of a libvirt error, a
LibvirtException is thrown.

Add processErrorIfZero(long) to ErrorHandler class to handle special
libvirt return codes, such as for virDomainGetMaxMemory. Use it in
Domain.getMaxMemory().

Signed-off-by: Claudio Bley <cbley av-test de>