Ensure JNA callbacks cannot get GCed
Currently nothing prevents the JNA callback objects used when registering
for domain events from being garbage collected.  JNA requires that callback
objects are not GCed whilst they are in use by C code.

To solve this we hold a reference to the callback alongside the callback id.
This ensures that the JNA callback objects will retain in memory while it is
registered with the C layer.

We also use an IdentityHashMap rather than a HashMap to store the EventListener
objects.